Moved the classes files into a classes/ folder.
