+++ all_render.php
@@ -2,7 +2,7 @@
     session_start();
     require_once ('FileMaker.php');
     require_once('utilities.php');
-    require_once ('DatabaseSearch.php');
+    require_once ('classes/DatabaseSearch.php');
     require_once ('partials/widgets.php');
     require_once ('credentials_controller.php');
