+++ classes/Specimen.php
