+++ search.php
@@ -5,8 +5,8 @@ use airmoi\FileMaker\Object\Field;
 
 require_once ('utilities.php');
 require_once ('constants.php');
-require_once ('DatabaseSearch.php');
-require_once ('Specimen.php');
+require_once ('classes/DatabaseSearch.php');
+require_once ('classes/Specimen.php');
 
 session_set_cookie_params(0,'/','.ubc.ca',isset($_SERVER["HTTPS"]), true);
 session_start();
